Include display event switch and export command to service graphs
