+++ www/front_src/src/Resources/Graph/Performance/ExportableGraphWithTimeline/exportToPng.ts
